better structure the code of the gorm endpoints and use gorm features more extensively for shorter and easier maintainable code
