+++ internal/database/models/projectpermission.go
@@ -2,6 +2,7 @@ package models
 
 import (
 	"database/sql/driver"
+	"fmt"
 
 	"github.com/google/uuid"
 )
@@ -15,7 +16,14 @@ const (
 )
 
 func (e *ProjectPermissionRole) Scan(value interface{}) error {
-	*e = ProjectPermissionRole(value.([]byte))
+	switch v := value.(type) {
+	case []byte:
+		*e = ProjectPermissionRole(v)
+	case string:
+		*e = ProjectPermissionRole(v)
+	default:
+		return fmt.Errorf("unsupported type: %T", v)
+	}
 	return nil
 }
 
@@ -26,6 +34,7 @@ func (e ProjectPermissionRole) Value() (driver.Value, error) {
 // ProjectPermission defines model for ProjectPermission.
 type ProjectPermission struct {
 	Role      ProjectPermissionRole `gorm:"type:ProjectPermissionRole"`
-	UserID    uuid.UUID
-	ProjectID uint32
+	UserID    uuid.UUID             `gorm:"primarykey"`
+	ProjectID uint32                `gorm:"primarykey"`
+	User      User
 }
