+++ internal/server/server-strict-gorm.go
@@ -2,14 +2,10 @@ package server
 
 import (
 	"context"
-	"fmt"
 	"miracummapper/internal/api"
 	"miracummapper/internal/config"
-	"miracummapper/internal/database/models"
 
-	"github.com/google/uuid"
 	"gorm.io/gorm"
-	"gorm.io/gorm/clause"
 )
 
 type StrictGormServer struct {
@@ -34,56 +30,6 @@ func (s *StrictGormServer) AddMapping(ctx context.Context, request api.AddMappin
 // AddPermission implements api.StrictServerInterface.
 func (s *StrictGormServer) AddPermission(ctx context.Context, request api.AddPermissionRequestObject) (api.AddPermissionResponseObject, error) {
 	panic("unimplemented")
-}
-
-// AddProject implements api.StrictServerInterface.
-func (s *StrictGormServer) AddProject(ctx context.Context, request api.AddProjectRequestObject) (api.AddProjectResponseObject, error) {
-	projectDetails := request.Body
-
-	// Validate the project details, must contain at least one code system role
-	if len(projectDetails.CodeSystemRoles) == 0 {
-		return api.AddProject422JSONResponse("CodeSystemRoles are required"), nil
-	}
-
-	// Create a new project
-	project := models.Project{
-		Name:                projectDetails.Name,
-		Description:         projectDetails.Description,
-		Version:             projectDetails.Version,
-		EquivalenceRequired: projectDetails.EquivalenceRequired,
-		StatusRequired:      projectDetails.StatusRequired,
-	}
-
-	// for i, role := range projectDetails.CodeSystemRoles {
-	// 	project.CodeSystemRoles = append(project.CodeSystemRoles, models.CodeSystemRole{
-	// 		Name:         role.Name,
-	// 		Type:         models.CodeSystemRoleType(role.Type),
-	// 		Position:     uint32(i),
-	// 		CodeSystemID: uint32(*role.System.Id),
-	// 	})
-	// }
-
-	for _, permission := range *projectDetails.ProjectPermissions {
-		userID, err := uuid.Parse(permission.UserId)
-		if err != nil {
-			return api.AddProject500JSONResponse{InternalServerErrorJSONResponse: "Invalid uuid provided"}, err
-		}
-		project.Permissions = append(project.Permissions, models.ProjectPermission{
-			Role:   models.ProjectPermissionRole(permission.Role),
-			UserID: userID,
-		})
-	}
-
-	// Create the project along with its associations
-	// s.Database.Clauses(clause.OnConflict{
-	// 	Columns:   []clause.Column{{Name: "project_id"}},
-	// 	DoUpdates: clause.AssignmentColumns([]string{"role", "user_id"}),
-	// }).Create(&project)
-
-	s.Database.Clauses(clause.OnConflict{DoNothing: true}).Create(&project)
-
-	// Return the ID of the newly created project
-	return api.AddProject200JSONResponse(*projectDetails), nil
 }
 
 // DeleteCodeSystem implements api.StrictServerInterface.
@@ -159,58 +105,6 @@ func (s *StrictGormServer) GetMappingByID(ctx context.Context, request api.GetMa
 // GetPermission implements api.StrictServerInterface.
 func (s *StrictGormServer) GetPermission(ctx context.Context, request api.GetPermissionRequestObject) (api.GetPermissionResponseObject, error) {
 	panic("unimplemented")
-}
-
-// GetProjects implements api.StrictServerInterface.
-func (s *StrictGormServer) GetProjects(ctx context.Context, request api.GetProjectsRequestObject) (api.GetProjectsResponseObject, error) {
-
-	pageSize := *request.Params.PageSize
-	offset := (*request.Params.Page - 1) * pageSize
-	sortBy := *request.Params.SortBy
-	switch sortBy {
-	case "dateCreated":
-		sortBy = "created"
-	}
-	sortOrder := *request.Params.SortOrder
-	switch sortOrder {
-	case "asc":
-		sortOrder = "ASC"
-	case "desc":
-		sortOrder = "DESC"
-	}
-
-	var projects []models.Project = []models.Project{}
-	err := s.Database.Order(fmt.Sprintf("%s %s", sortBy, sortOrder)).Offset(offset).Limit(pageSize).Find(&projects).Error
-	if err != nil {
-		return nil, err
-	}
-
-	// Convert projects to api.Projects
-	var apiProjects []api.Project = []api.Project{}
-	for _, project := range projects {
-		apiProjects = append(apiProjects, convertToAPIProject(project))
-	}
-
-	return api.GetProjects200JSONResponse(apiProjects), nil
-}
-
-func convertToAPIProject(project models.Project) api.Project {
-	id := int32(project.ID)
-	var modified string
-	if !project.UpdatedAt.IsZero() {
-		modified = project.UpdatedAt.String()
-	} else {
-		modified = ""
-	}
-	return api.Project{
-		Description:         project.Description,
-		EquivalenceRequired: project.EquivalenceRequired,
-		Id:                  &id,
-		Modified:            &modified, // Assign the string pointer
-		Name:                project.Name,
-		StatusRequired:      project.StatusRequired,
-		Version:             project.Version,
-	}
 }
 
 // // Ping implements api.StrictServerInterface.
