+++ internal/utilities/requests.go
@@ -1,0 +1,5 @@
+package utilities
+
+func GetOffset(page int, pageSize int) int {
+	return (page - 1) * pageSize
+}
